Update README.md
This is the snake game, and it is fun to play. The user controls the direction of the creature and every this it eats the food, its length increases. The score keeps updating simultaneously, and the game gets over when it hits itself. Now you can either restart or end the game.

Tech Stacks - IntelliJ Software, Java Swing GUI